Add an test to ensure that local manifest file matches the live manifest file
