--- scripts/ampHtmlValidator/cli.js
@@ -1,3 +1,0 @@
-const { runValidator } = require('.');
-
-runValidator();
